+++ src/collection/SubCollectionsView.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Box, Button, Grid, Paper, Tab, Tabs } from "@material-ui/core";
+import { Box, Button, Grid, Paper, Tab, Tabs, Typography } from "@material-ui/core";
 import { Entity, EntityCollectionView, EntitySchema } from "../models";
 import { Link as ReactLink } from "react-router-dom";
 import {
@@ -26,6 +26,7 @@ interface SubCollectionViewProps<S extends EntitySchema> {
                   entity: Entity<any>,
                   schema: EntitySchema<any>,
                   subcollections?: EntityCollectionView<any>[]): void;
+
 }
 
 export default function SubCollectionsView<S extends EntitySchema>(
@@ -45,6 +46,11 @@ export default function SubCollectionsView<S extends EntitySchema>(
 
     const firstSubcollection = Object.values(subcollections)[0];
     const [selectedView, setSelectedView] = React.useState<EntityCollectionView<any>>(firstSubcollection);
+    const [deleteEntityClicked, setDeleteEntityClicked] = React.useState<Entity<S> | undefined>(undefined);
+
+    const onEntityDelete = (collectionPath: string, entity: Entity<S>) => {
+        setDeleteEntityClicked(entity);
+    };
 
     function _onViewClicked(view: EntityCollectionView<any>) {
         setSelectedView(view);
@@ -59,6 +65,7 @@ export default function SubCollectionsView<S extends EntitySchema>(
         <Paper elevation={0}
                className={classes.formPaper}
                style={{ height: "100%" }}>
+
             <Grid
                 container
                 direction="row"
@@ -72,17 +79,6 @@ export default function SubCollectionsView<S extends EntitySchema>(
                              key={`wrapped-tab-${key}`}/>
                     ))}
                 </Tabs>
-                <Box my={2} textAlign="right">
-                    {subcollectionPath && <Button
-                        component={ReactLink}
-                        to={getRouterNewEntityPath(subcollectionPath)}
-                        size="medium"
-                        variant="outlined"
-                        color="primary"
-                    >
-                        Add {selectedView.schema.name}
-                    </Button>}
-                </Box>
             </Grid>
 
             {Object.entries(subcollections).map(([key, view]) => {
@@ -94,75 +90,63 @@ export default function SubCollectionsView<S extends EntitySchema>(
                     if (onEntityEdit)
                         onEntityEdit(collectionPath, entity, view.schema, view.subcollections);
                 };
+
+                const title =
+                    <Typography variant={"caption"} color={"textSecondary"}>
+                        {`/${subcollectionPath}`}
+                    </Typography>;
+
+                const deleteEnabled = view.deleteEnabled === undefined || view.deleteEnabled;
+
                 return (
-                    <SubCollectionTabPanel
-                        key={`wrapped-tab-content-${key}`}
-                        small={view.small === undefined ? false : view.small}
-                        selectedView={selectedView}
-                        thisView={view}
-                        onEntityEdit={onSubcollectionEntityEdit}
-                        onEntityClick={onSubcollectionEntityClick}
-                        subcollectionPath={subcollectionPath}/>
+                <Box
+                    style={{ height: "500px" }}
+                    hidden={selectedView !== view}>
+
+                    {subcollectionPath ?
+                        <CollectionTable collectionPath={subcollectionPath}
+                                         title={title}
+                                         onEntityDelete={deleteEnabled ? onEntityDelete : undefined}
+                                         schema={view.schema}
+                                         onEntityEdit={onSubcollectionEntityEdit}
+                                         onEntityClick={onSubcollectionEntityClick}
+                                         includeToolbar={true}
+                                         paginationEnabled={false}
+                                         additionalColumns={view.additionalColumns}
+                                         defaultSize={view.defaultSize}
+                                         actions={
+                                             <Button
+                                                 component={ReactLink}
+                                                 to={getRouterNewEntityPath(subcollectionPath)}
+                                                 size="medium"
+                                                 variant="outlined"
+                                                 color="primary"
+                                             >
+                                                 Add {selectedView.schema.name}
+                                             </Button>
+                                         }
+                        />
+                        :
+                        <Box m={3}
+                             display={"flex"}
+                             alignItems={"center"}
+                             justifyContent={"center"}>
+                            <Box>
+                                You need to save your entity before adding
+                                additional collections
+                            </Box>
+                        </Box>
+                    }
+
+                    <DeleteEntityDialog entity={deleteEntityClicked}
+                                        schema={view.schema}
+                                        open={!!deleteEntityClicked}
+                                        onClose={() => setDeleteEntityClicked(undefined)}/>
+                </Box>
                 );
             })}
+
         </Paper>
     );
 }
 
-
-export interface SubCollectionTabPanelProps<S extends EntitySchema> {
-    subcollectionPath: string | undefined;
-    small: boolean;
-    thisView: EntityCollectionView<S>;
-    selectedView: EntityCollectionView<S>;
-
-    onEntityEdit?(collectionPath: string, entity: Entity<any>): void;
-
-    onEntityClick?(collectionPath: string, entity: Entity<any>): void;
-}
-
-export function SubCollectionTabPanel<S extends EntitySchema>({ subcollectionPath, selectedView, thisView, onEntityEdit, onEntityClick, small, ...props }: SubCollectionTabPanelProps<S>) {
-
-    const [deleteEntityClicked, setDeleteEntityClicked] = React.useState<Entity<S> | undefined>(undefined);
-
-    const onEntityDelete = (collectionPath: string, entity: Entity<S>) => {
-        setDeleteEntityClicked(entity);
-    };
-
-    const deleteEnabled = thisView.deleteEnabled === undefined || thisView.deleteEnabled;
-
-    return (
-        <Box
-            style={{height: "400px"}}
-            hidden={selectedView !== thisView}>
-
-            {subcollectionPath ?
-                <CollectionTable collectionPath={subcollectionPath}
-                                 onEntityDelete={deleteEnabled ? onEntityDelete : undefined}
-                                 schema={thisView.schema}
-                                 onEntityEdit={onEntityEdit}
-                                 onEntityClick={onEntityClick}
-                                 includeToolbar={false}
-                                 paginationEnabled={false}
-                                 small={small}
-                                 additionalColumns={thisView.additionalColumns}
-                />
-                :
-                <Box m={3}
-                     display={"flex"}
-                     alignItems={"center"}
-                     justifyContent={"center"}>
-                    <Box>
-                        You need to save your entity before adding
-                        additional collections
-                    </Box>
-                </Box>
-            }
-
-            <DeleteEntityDialog entity={deleteEntityClicked}
-                                schema={thisView.schema}
-                                open={!!deleteEntityClicked}
-                                onClose={() => setDeleteEntityClicked(undefined)}/>
-        </Box>
-    );
-}